flask-hello-world-app/flask-hello-world-app/requirements.txt
flask
databricks-sdk
databricks-sql-connector
pandas
pyarrow
PyPDF2
pdfplumber
openai
groq
